move pipeline from list
